+++ neural_process_model.py
@@ -7,13 +7,6 @@ tfk = tf.keras
 """https://github.com/revsic/tf-neural-process/blob/master/neural_process/"""
 
 def dense_sequential(output_sizes, activation=tf.nn.relu):
-    """Convert number of hidden units to sequential dense layers
-    Args:
-        output_sizes: List[int], number of hidden units
-        activation: Callable[[tf.Tensor], tf.Tensor], activation function, default ReLU
-    Returns:
-        tfk.Model, sequential model consists of dense layers
-    """
     model = tfk.Sequential()
     for size in output_sizes[:-1]:
         model.add(tfk.layers.Dense(size, activation=activation))
@@ -22,13 +15,6 @@ def dense_sequential(output_sizes, activation=tf.nn.relu):
     return model
 
 class Encoder(tfk.layers.Layer):
-    """Context encoder
-    Attributes:
-        model: Callable[[tf.Tensor], tf.Tensor], dense sequential encoder
-        attention: Callable[[tf.Tensor], tf.Tensor], attention method, default None
-            if None, attention is not applied
-        keepdims: bool, if false, reduce axis 1 with mean method
-    """
     def __init__(self, output_sizes, attention=None, keepdims=False, name='Encoder'):
         super(Encoder, self).__init__(name=name, dynamic=True)
         self.model = dense_sequential(output_sizes)
@@ -36,15 +22,6 @@ class Encoder(tfk.layers.Layer):
         self.keepdims = keepdims
 
     def call(self, rep, key=None, query=None):
-        """Encode given context
-        Args:
-            rep: tf.Tensor or tuple, list of tf.Tensor, representation
-                if rep consists of multiple tensor, concatenate it to decode
-            key: tf.Tensor, key for attention method, default None
-            query: tf.Tensor, query for attention method, default None
-        Returns:
-            tf.Tensor, encoded context
-        """
         if isinstance(rep, (tuple, list)):
             rep = tf.concat(rep, axis=-1)
 
@@ -59,100 +36,53 @@ class Encoder(tfk.layers.Layer):
 
 
 class Decoder(tfk.layers.Layer):
-    """Context decoder
-    Attributes:
-        model: Callable[[tf.Tensor], tf.Tensor], dense sequential decoder
-    """
     def __init__(self, output_sizes, name='Decoder'):
         super(Decoder, self).__init__(name=name, dynamic=True)
         self.model = dense_sequential(output_sizes)
     
     def call(self, context, tx):
-        """Decode tensor
-        Args:
-            context: tf.Tensor, encoded context
-            tx: tf.Tesnor, query
-        
-        Returns:
-            tf.Tensor, decoded value
-        """
         input_tensor = tf.concat([context, tx], axis=-1)
         return self.model(input_tensor)
 
 
-class GaussianProb(tfk.layers.Layer):
-    """Convert input tensor to gaussian distribution representation
-    Attributs:
-        dense_mu: Callable[[tf.Tensor], tf.Tensor], dense layer for mean
-        dense_sigma: Callable[[tf.Tensor], tf.Tensor], dense layer for sigma
-        multivariate: bool, if true, return multivariate gaussian distribution with diagonal covariance
-        proj: Callable[[tf.Tensor], tf.Tensor], projection layer for input tensor, default None
-            if None, projection is not applied
-    """
-    def __init__(self, size, multivariate=False, proj=None, name='GP'):
-        super(GaussianProb, self).__init__(name=name, dynamic=True)
-        self.dense_mu = tfk.layers.Dense(size)
-        self.dense_sigma = tfk.layers.Dense(size)
-        self.multivariate = multivariate
+class LatentEncoder(tfk.layers.Layer):
+    def __init__(self, output_sizes, name='LatentVariable'):
+        super(LatentEncoder, self).__init__(name=name, dynamic=True)
+        self.model = dense_sequential(output_sizes)
 
-        self.proj = proj
-        if proj is not None:
-            self.proj = tfk.layers.Dense(proj)
-    
-    def call(self, input_tensor):
-        if self.proj is not None:
-            input_tensor = self.proj(input_tensor)
-        
-        mu = self.dense_mu(input_tensor)
-        log_sigma = self.dense_sigma(input_tensor)
-        
+    def call(self, rep):
+        hidden = self.model(rep)
+        hidden = tf.reduce_mean(hidden, axis=1) # the first axis is the number of samples, remove these, LOOSING INFORMATION
+
+        mu, log_sigma = tf.split(hidden, num_or_size_splits=2, axis=-1) # split the output in half
+
         sigma = tf.exp(log_sigma)
-        if self.multivariate:
-            dist = tfp.distributions.MultivariateNormalDiag(mu, sigma)
-        else:
-            dist = tfp.distributions.Normal(mu, sigma)
-        
-        return dist, mu, sigma 
+        dist = tfp.distributions.Normal(mu, sigma)
+        return dist
+
 
 class NeuralProcess(tfk.Model):
-    """Neural Process
-    Attributes:
-        z_encoder: Encoder, encoder for latent representation
-        z_prob: GaussianProb, latent representation to probability distribution
-        encoder: Encoder, context encoder
-        decoder: Decoder, decoder for context and latent variable
-        normal_dist: GaussianProb, converter for decoded context to probability distribution
-    """
     def __init__(self,
                  z_output_sizes,
                  enc_output_sizes,
-                 dec_output_sizes, name='Neural Process'):
-        """Initializer
-        Args:
-            z_output_sizes: List[int], number of hidden units for latent representation encoder
-            enc_output_sizes: List[int], number of hidden units for context encoder
-            dec_output_sizes: List[int], number of hidden units for decoder
-        """
+                 dec_output_sizes, name='NeuralProcess'):
         super(NeuralProcess, self).__init__(name=name)
 
-        self.z_encoder = Encoder(z_output_sizes[:-1])
-        self.z_prob = GaussianProb(z_output_sizes[-1],
-                                   proj=np.mean(z_output_sizes[-2:]))
-
+        self.z_encoder_latent = LatentEncoder(z_output_sizes)
         self.encoder = Encoder(enc_output_sizes)
-        self.decoder = Decoder(dec_output_sizes[:-1])
-        self.normal_dist = GaussianProb(dec_output_sizes[-1], multivariate=True)
+        self.decoder = Decoder(dec_output_sizes)#[:-1])
+
+        # size = dec_output_sizes[-1]
+        # self.dense_mu = tfk.layers.Dense(size)
+        # self.dense_sigma = tfk.layers.Dense(size)
     
     @tf.function
-    def call(self, inputs):
-        context_x, context_y, target_x = inputs # pixel coords, pixel values, all image coords, target image
-        #context = context_y # context pixel values
-        context = tf.concat([context_x, context_y], axis=-1)
-        query = target_x # the pixel coords we want to predict - all image coords
-        z_context = self.z_encoder(context)
-        z_dist, _, _ = self.z_prob(z_context)
+    def call(self, x):
+        context, query = x
 
+        z_dist = self.z_encoder_latent(context)
         latent = z_dist.sample()
+        
         context = self.encoder(context)
 
         context = tf.concat([latent, context], axis=-1)
@@ -160,21 +90,10 @@ class NeuralProcess(tfk.Model):
                           [1, tf.shape(query)[1], 1])
 
         rep = self.decoder(context, query)
-        dist, mu, sigma = self.normal_dist(rep)
+        mu, log_sigma = tf.split(rep, num_or_size_splits=2, axis=-1) # split the output in half
+        # mu = self.dense_mu(rep)
+        # log_sigma = self.dense_sigma(rep)
+        
+        sigma = tf.exp(log_sigma)
 
-        return tf.concat([mu, sigma], axis=-1) #dist, mu, sigma
-    
-    # def loss(self, context, query, target):
-    #     dist, _, _ = self(context, query)
-    #     log_prob = dist.log_prob(target)
-    #     log_prob = tf.reduce_sum(log_prob)
-
-    #     prior, _, _ = self.z_prob(self.z_encoder(context))
-    #     posterior, _, _ = self.z_prob(self.z_encoder([query, target]))
-
-    #     kl = tfp.distributions.kl_divergence(prior, posterior)
-    #     kl = tf.reduce_sum(kl)
-
-    #     # maximize variational lower bound
-    #     loss = -log_prob + kl
-    #     return loss
+        return tf.concat([mu, sigma], axis=-1)#(dist, mu, sigma) # tf.concat([mu, sigma], axis=-1) #dist, mu, sigma
